+++ tests/Pest.php
@@ -11,6 +11,7 @@
 |
 */
 
+use LogScope\Tests\ApiEnabledTestCase;
 use LogScope\Tests\AutoScheduleEnabledTestCase;
 use LogScope\Tests\EagerProviderTestCase;
 use LogScope\Tests\TestCase;
@@ -20,3 +21,4 @@ pest()->extend(TestCase::class)->in('Feature', 'Unit');
 pest()->extend(TransactionTestCase::class)->in('Transactions');
 pest()->extend(EagerProviderTestCase::class)->in('EagerProvider');
 pest()->extend(AutoScheduleEnabledTestCase::class)->in('AutoSchedule');
+pest()->extend(ApiEnabledTestCase::class)->in('Api');